refactor: split Excalidraw socket managers
